App/requirements.txt
fastapi
uvicorn
scikit-learn
joblib
psycopg2-binary
pandas